DATT 3701 Project
Prison Escape is a 3v1 local multiplayer party game. 3 players will take on the role of monkeys and work together to escape a dangerous laboratory.  1 player will take on the role of a scientist attempting to prevent the monkeys from escaping the laboratory.

Prison Escape is fun for all and its quick 10 minute rounds will leave players wanting more.